# Complete refactor around "acorn" and "escope"