--- a/package.json
+++ b/package.json
@@ -32,7 +32,6 @@
     "yargs": "^1.3.1"
   },
   "devDependencies": {
-    "chai": "~1.8.1",
     "mocha": "~1.16.2",
     "istanbul": "~0.2.1"
   }