--- a/package.json
+++ b/package.json
@@ -29,10 +29,10 @@
     "esprima-fb": "^10001.1.0-dev-harmony-fb",
     "esrefactor": "~0.1.0",
     "estraverse": "~1.9.1",
-    "yargs": "^1.3.1"
+    "yargs": "^3.4.5"
   },
   "devDependencies": {
-    "mocha": "~1.16.2",
-    "istanbul": "~0.2.1"
+    "mocha": "^2.2.0",
+    "istanbul": "^0.3.7"
   }
 }