… test/ember-data.dereq.js
… test/react-with-addons.dereq.js
… test/react-with-addons.js