This is a much faster and simpler implementation of `derequire`. 

The only "breaking" change is the way it handled `test/cjs-smartass.js`. I'm not exactly sure what it was doing before, but the new behavior seems more correct.

There is now a perf test that can be started with `npm run perf`. It `derequires` React's `react-with-addons.js` for two tokens. Here are the results:

**Before:**
```
derequire "require" and "element"
---------------------------------
1/5 took 8982ms
2/5 took 8987ms
3/5 took 9635ms
4/5 took 9606ms
5/5 took 9726ms
avg 9387ms
```

**After:**
```
derequire "require" and "element"
---------------------------------
1/5 took 327ms
2/5 took 243ms
3/5 took 223ms
4/5 took 214ms
5/5 took 207ms
avg 242ms
```

That's a 97% decrease in run time, or a 38,700% speed increase :smile: