@calvinmetcalf 2015-03-07
It looks good, I think cjs-lazy was purely a perf thing ie don't parse the file if there are no calls to require. Also feel free to add the bench mark.