@calvinmetcalf 2015-03-07
Though on second thought the fact we are switching to acorn means it'll
have to be a major version bump anyway so the lazy thing should be fine

On Sat, Mar 7, 2015, 1:32 AM Andres Suarez <notifications@github.com> wrote:

> The tests took around ~780ms, now they take ~300ms.
>
> Results from running derequire on react-with-addons.js
>
> # before
> ❯ time cat ../react-with-addons.js | ./bin/cmd.js > /dev/null
>
> real    0m7.235s
> user    0m7.194s
> sys 0m0.074s
> # after
> ❯ time cat ../react-with-addons.js | ./bin/cmd.js > /dev/null
>
> real    0m0.432s
> user    0m0.389s
> sys 0m0.046s
>
> I disagree with the way test/cjs-lazy.js was being treated, but I'm open
> to going back.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/calvinmetcalf/derequire/pull/26
> Commit Summary
>
>    - Total refactor around "acorn" and "escope"
>
> File Changes
>
>    - *M* index.js
>    <https://github.com/calvinmetcalf/derequire/pull/26/files#diff-0>
>    (123)
>    - *M* package.json
>    <https://github.com/calvinmetcalf/derequire/pull/26/files#diff-1> (5)
>    - *M* test/cjs-lazy.js
>    <https://github.com/calvinmetcalf/derequire/pull/26/files#diff-2> (4)
>
> Patch Links:
>
>    - https://github.com/calvinmetcalf/derequire/pull/26.patch
>    - https://github.com/calvinmetcalf/derequire/pull/26.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/calvinmetcalf/derequire/pull/26>.
>