@zertosh 2015-03-07
@calvinmetcalf Sounds good. I've still got to add a few more tests, a benchmark, and a few more optimizations. I'll get around to it later today - I'd love to see this out by Monday.